Revert "Restructure and protect routes (#24)"
This reverts commit 1238963bb7dc25562ac9fe4a9ee1af4eb8dadc41.
