+++ client/src/.env.example
@@ -1,0 +1,2 @@
+REACT_APP_AUTH0_DOMAIN=
+REACT_APP_CLIENT_ID=
