+++ client/src/Pages/Experts.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+class Experts extends React.Component {
+    render() {
+      return (
+        <div>
+          <h1>Experts</h1>
+        </div>
+      )
+    }
+  }
+export default Experts;
