+++ client/src/Pages/Landing.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+class Landing extends React.Component {
+    render() {
+      return (
+        <div>
+          <h1>Landing</h1>
+        </div>
+      )
+    }
+  }
+export default Landing;
