+++ client/src/Pages/Learners.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Learners extends React.Component {
+    render() {
+      return (
+        <div>
+          <h1>Learners</h1>
+        </div>
+      )
+    }
+  }
+
+export default Learners;
