+++ client/src/Pages/Profile.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Profile extends React.Component {
+  render() {
+    return (
+      <div>
+        <h1>Profile</h1>
+      </div>
+    )
+  }
+}
+
+export default Profile;
