+++ client/src/index.js
@@ -1,21 +1,19 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { BrowserRouter } from 'react-router-dom'
 import { Auth0Provider } from "@auth0/auth0-react";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 
+// TODO: use react router :)
 ReactDOM.render(
   <React.StrictMode>
-    <BrowserRouter>
-      <Auth0Provider
-        domain={process.env.REACT_APP_AUTH0_DOMAIN}
-        clientId={process.env.REACT_APP_CLIENT_ID}
-        redirectUri={window.location.origin}
-      >
-        <App />
-      </Auth0Provider>
-    </BrowserRouter>
+    <Auth0Provider
+      domain={process.env.REACT_APP_AUTH0_DOMAIN}
+      clientId={process.env.REACT_APP_CLIENT_ID}
+      redirectUri={window.location.origin}
+    >
+      <App />
+    </Auth0Provider>
   </React.StrictMode>,
   document.getElementById("root")
 );
