--- client/src/layout/Footer/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from './index.module.css'
-
-const Footer = () => {
-  return (
-    <footer className={styles.footer}>
-      Footer
-    </footer>
-  )
-}
-
-export default Footer;
