--- client/src/layout/Footer/index.module.css
@@ -1,7 +1,0 @@
-footer {
-  height: 60px;
-  width: 100vw;
-  position: absolute;
-  bottom: 0;
-  background-color: #fafafa;
-}
