--- client/src/pages/Experts/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from './index.module.css'
-
-const Experts = () => {
-  return (
-    <>
-      <h1>Experts</h1>
-    </>
-  )
-}
-
-export default Experts;
