--- client/src/pages/Fallback/index.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import styles from './index.module.css'
-
-const Fallback = () => {
-  return (
-    <h1>Page Not Found</h1>
-  )
-}
-
-export default Fallback;
