--- client/src/pages/Landing/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from './index.module.css'
-
-const Landing = () => {
-  return (
-    <>
-      <h1>Landing</h1>
-    </>
-  )
-}
-
-export default Landing;
