--- client/src/pages/Learners/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from './index.module.css'
-
-const Learners = () => {
-  return (
-    <>
-      <h1>Learners</h1>
-    </>
-  )
-}
-
-export default Learners;
