--- client/src/pages/Profile/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from './index.module.css'
-
-const Profile = () => {
-  return (
-    <>
-      <h1>Profile</h1>
-    </>
-  )
-}
-
-export default Profile;
